Updated game options
I added an option for the user to choose if they want to play best out of 3, 5,or 7. When the user is done playing, the console will output how many games they won and how many total rounds were played. I also updated the inputs from the user so it won't move on the code until it has a valid input.